move hoh modules
