+++ BossMod/Modules/Stormblood/DeepDungeon/HeavenOnHigh/D50Gozu.cs
